+++ advanced_scripts/photo_collection.sh
@@ -16,6 +16,8 @@ PHOTO_DURATION=2
 MAX_PHOTO_ANGLE=25
 BACKGROUND_COLOR="#00000000"
 
+IFS=$'\t\n'                 # NECESSARY TO SUPPORT SPACE IN FILE NAMES
+
 # PHOTO OPTIONS - ALL FILES UNDER photos FOLDER ARE USED - USE sort TO SPECIFY A SORTING MECHANISM
 # PHOTOS=`find ../photos/* | sort -r`
 PHOTOS=`find ../photos/*`
@@ -49,7 +51,7 @@ FULL_SCRIPT="ffmpeg -y "
 
 # 2. ADD INPUTS
 for photo in ${PHOTOS}; do
-    FULL_SCRIPT+="-loop 1 -i ${photo} "
+    FULL_SCRIPT+="-loop 1 -i '${photo}' "
 done
 
 # 3. ADD BACKGROUND COLOR SCREEN INPUT
@@ -85,3 +87,5 @@ eval ${FULL_SCRIPT}
 ELAPSED_TIME=$(($SECONDS - $START_TIME))
 
 echo -e '\nSlideshow created in '$ELAPSED_TIME' seconds\n'
+
+unset $IFS