+++ transition_scripts/stack_vertical.sh
@@ -17,6 +17,8 @@ DIRECTION=2                 # 1=TOP TO BOTTOM, 2=BOTTOM TO TOP
 INCLUDE_INTRO=1             # START WITH EMPTY SCREEN
 INCLUDE_OUTRO=0             # END WITH EMPTY SCREEN
 
+IFS=$'\t\n'                 # NECESSARY TO SUPPORT SPACE IN FILE NAMES
+
 # PHOTO OPTIONS - ALL FILES UNDER photos FOLDER ARE USED - USE sort TO SPECIFY A SORTING MECHANISM
 # PHOTOS=`find ../photos/* | sort -r`
 PHOTOS=`find ../photos/*`
@@ -43,7 +45,7 @@ FULL_SCRIPT="ffmpeg -y "
 
 # 2. ADD INPUTS
 for photo in ${PHOTOS}; do
-    FULL_SCRIPT+="-loop 1 -i ${photo} "
+    FULL_SCRIPT+="-loop 1 -i '${photo}' "
 done
 
 # 3. ADD BACKGROUND COLOR SCREEN INPUT
@@ -98,3 +100,5 @@ eval ${FULL_SCRIPT}
 ELAPSED_TIME=$(($SECONDS - $START_TIME))
 
 echo -e '\nSlideshow created in '$ELAPSED_TIME' seconds\n'
+
+unset $IFS